separated pages and screens for bus_checkin_roster, updated bus_routes
